features
rewrite api with basic auth and isAdmin middleware
added 403 page with automatic redirection after 10s timeout
in API get stream added charts data fro each pid

+++ routes/api.php
@@ -1,26 +1,29 @@
 <?php
 
-use App\Http\Controllers\API\ApiStreamAnalyzeController;
-use App\Http\Controllers\API\ApiStreamController;
-use App\Http\Controllers\Streams\ApiStreamAlertController;
-use App\Http\Controllers\Streams\ApiStreamImageController;
-use App\Http\Controllers\Streams\ApiStreamStopController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\API\ApiStreamController;
+use App\Http\Controllers\API\ApiStreamAnalyzeController;
+use App\Http\Controllers\Streams\ApiStreamStopController;
+use App\Http\Controllers\Streams\ApiStreamAlertController;
+use App\Http\Controllers\Streams\ApiStreamImageController;
+use App\Http\Controllers\Streams\ApiStreamStartController;
 
 Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
     return $request->user();
 });
 
-Route::prefix('v2')->group(function () {
-    Route::prefix('stream')->group(function () {
-        Route::get('alerts', ApiStreamAlertController::class);
-        Route::post('', [ApiStreamController::class, 'show']);
-        Route::post('create', [ApiStreamController::class, 'store']);
-        Route::delete('{stream}', [ApiStreamController::class, 'destroy']);
-        Route::post('analyze', ApiStreamAnalyzeController::class);
-        Route::get('image/{stream}', ApiStreamImageController::class);
-        Route::patch('{stream}/stop', ApiStreamStopController::class);
-        Route::patch('{stream}/start', ApiStreamStopController::class);
+Route::middleware('auth.basic')->group(function () {
+    Route::prefix('v2')->group(function () {
+        Route::prefix('stream')->group(function () {
+            Route::get('alerts', ApiStreamAlertController::class);
+            Route::get('{stream}', [ApiStreamController::class, 'show'])->middleware('isAdmin');
+            Route::post('', [ApiStreamController::class, 'store'])->middleware('isAdmin');
+            Route::delete('{stream}', [ApiStreamController::class, 'destroy'])->middleware('isAdmin');
+            Route::post('analyze', ApiStreamAnalyzeController::class)->middleware('isAdmin');
+            Route::get('{stream}/image', ApiStreamImageController::class)->middleware('isAdmin');
+            Route::patch('{stream}/stop', ApiStreamStopController::class)->middleware('isAdmin');
+            Route::patch('{stream}/start', ApiStreamStartController::class)->middleware('isAdmin');
+        });
     });
 });
